he corregido el error del casteo de int a double en automovil, la lo puede registrar, validen que no ingrese un enter y sale error cuando pongo opcion 9

+++ src/UI.java
@@ -107,7 +107,7 @@ public class UI {
         System.out.print("Ingrese el propietario: ");
         String propietario = scanner.nextLine();
         datos.add(propietario);
-        int cilindraje = (int) validar.ingresarValorNumerico(0, "Ingrese el cilindraje (en cm3): ");
+        double cilindraje =  validar.ingresarValorNumerico(0, "Ingrese el cilindraje (en cm3): ");
         datos.add(cilindraje);
         double precio_sugerido = validar.ingresarValorNumerico(0, "Ingrese el valor sugerido: ");
         datos.add(precio_sugerido);
